feat: Update customWebName and virtualNetwork modules
Bug fix: The code changes in the `customWebName.bicep` module update the `subdomain` parameter to use the `subdomain` variable instead of a hardcoded value. This change allows for more flexibility in configuring the subdomain for the custom web name, whilst fixing a bug because the full fqdn was used by mistake.

Bug fix: In the `virtualNetwork.bicep` module, the `natGateway` property under `gatewaySubnetName` has been removed. This change removes the dependency on a NAT gateway for the gateway subnet that prevented its deployment.

These updates improve the configuration and flexibility of the OCPP server infrastructure.