feat: Update User Assigned Managed Identity permissions for SSL certificate access
The code changes in the README.md file update the instructions for setting up the OCPP server. Specifically, it adds information about setting Get and List certificate permissions in the certificate access policy for the User Assigned Managed Identity. This ensures that the SSL certificate stored in the KeyVault can be accessed by the OCPP server.