feat: Update appGw to conditionally set the host property
The code changes in the appgw.bicep file update the host property of the appGw resource. It now checks the value of the isSecure variable and sets the host property accordingly. If isSecure is true, it uses the webHostName variable. Otherwise, it uses the defaultHostName property of the webApp resource. This change allows for more flexibility in configuring the host property based on the presence of a domain name.