feat: Add testvars target to Makefile for retrieving test variables
The code changes in the Makefile and webapp.mk files add a new target called "testvars" to retrieve test variables. This target allows developers to easily retrieve and display the test variables needed for running the OCPP server tests. It improves the development workflow by providing a convenient way to access the necessary test configuration.