Jmservera/conditionaldeployment (#19)
* set appgw conditional deployment depending on having a domain name or not.

* feat: Update appGw to conditionally set the host property

The code changes in the appgw.bicep file update the host property of the appGw resource. It now checks the value of the isSecure variable and sets the host property accordingly. If isSecure is true, it uses the webHostName variable. Otherwise, it uses the defaultHostName property of the webApp resource. This change allows for more flexibility in configuring the host property based on the presence of a domain name.

* feat: Update Makefile to improve operation. Added a child makefile for the web app publishing that needs the resources to be already created in Azure. This allows to run the main makefile without having to retrieve the values with the az command when they are not needed

* feat: Add testvars target to Makefile for retrieving test variables

The code changes in the Makefile and webapp.mk files add a new target called "testvars" to retrieve test variables. This target allows developers to easily retrieve and display the test variables needed for running the OCPP server tests. It improves the development workflow by providing a convenient way to access the necessary test configuration.

* fix: npm vulnerabilities

* feat: Add support for User Assigned Managed Identity in web app deployment

The code changes in the main.bicep and webapp.bicep files add support for User Assigned Managed Identity in the deployment of the web app. This allows the web app to access the SSL certificate stored in the KeyVault.

* fix typo

* add flag for custom cert in internal app